+++ extensions/amazon-aws/src/ec2.tsx
@@ -6,7 +6,7 @@ import { useCachedPromise } from "@raycast/utils";
 const preferences = setupAws();
 const ec2 = new AWS.EC2({ apiVersion: "2016-11-15" });
 
-export default function DescribeInstances() {
+export default function EC2() {
   const { data: instances, error, isLoading } = useCachedPromise(fetchEC2Instances);
 
   if (error) {
@@ -18,13 +18,13 @@ export default function DescribeInstances() {
   return (
     <List isLoading={isLoading} searchBarPlaceholder="Filter instances by name...">
       {instances?.map((i) => (
-        <InstanceListItem key={i.InstanceId} instance={i} />
+        <EC2Instance key={i.InstanceId} instance={i} />
       ))}
     </List>
   );
 }
 
-function InstanceListItem(props: { instance: AWS.EC2.Instance }) {
+function EC2Instance(props: { instance: AWS.EC2.Instance }) {
   const instance = props.instance;
   const name = instance.Tags?.find((t) => t.Key === "Name")?.Value?.replace(/-/g, " ");
 
@@ -57,7 +57,7 @@ function InstanceListItem(props: { instance: AWS.EC2.Instance }) {
       key={instance.InstanceId}
       title={name || "Unknown Instance name"}
       subtitle={instance.InstanceType}
-      icon="list-icon.png"
+      icon="ec2.png"
       actions={
         <ActionPanel>
           <Action.OpenInBrowser